requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
django-environ==0.10.0
django-filter==23.2
environ==1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pyspark==3.2.0
sqlparse==0.4.4
typing-extensions==4.6.3
